requirements.txt
fastapi
uvicorn
opencv-python-headless
python-multipart
requests
jinja2
aiofiles
psutil
pydantic
pydantic-settings
numpy
PyYAML
itsdangerous
